% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lee_img.R
\name{carga_data}
\alias{carga_data}
\title{Lee data Imgenes
xD}
\usage{
carga_data(dirr, canal)
}
\arguments{
\item{dirr}{Directorio}

\item{canal}{Canal de imagen "red" , "green", "blue"}
}
\value{
\code{Data Science}.
}
\description{
Lee data Imgenes
xD
}
\examples{
##-----Ejemplito------
  data <- carga_data("E:/LuisxM/Desktop/idc", "red")
}
\author{
\strong{Luis F. Maron}
}
