% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafica.R
\name{grafica}
\alias{grafica}
\title{Grafica de Funcion
xD}
\usage{
grafica(fun, a, b)
}
\arguments{
\item{fun}{Funcion}

\item{a}{Limite inicial, izquierdo}

\item{b}{Limite final, derecho}
}
\value{
\code{Grafiquito}.
}
\description{
Grafica de Funcion
xD
}
\examples{
##-----Ejemplito------
 fn <- function(x) x**3-5*x
 grafica(fn, -5, 5)
}
\author{
\strong{Luis F. Maron}
}
