\name{A. Metodos Numericos:}
\alias{A) Metodos Numericos:}
\encoding{utf-8}
\title{ \deqn{Paquete Métodos Numéricos}}
\details{
Package:    metodNum \cr
Version:	   1.0.2 \cr
Date:       2018-01-19 \cr
License:	   GPL (>=2) \cr
Packaged:   2017-05-09 10:31:35 UTC; root \cr
Built:	   R 3.4.3; ; 2017-05-09 20:45:26 UTC; linux

}
\author{
  → \strong{Luis Fernando MARON I.} \cr
  → <\email{luisfer4222@mail.com}> \cr
  → \pkg{\code{\href{https://www.facebook.com/luis.f.maron}{My Facebook}}}
}
