% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacios.R
\name{vaciosxD}
\alias{vaciosxD}
\title{Data Vacia missing
xD}
\usage{
vaciosxD(data, por, plot)
}
\arguments{
\item{data}{Data}

\item{por}{Porcentaje de decision}

\item{plot}{Muestra grafico}
}
\value{
\code{Data Science}.
}
\description{
Data Vacia missing
xD
}
\examples{
##-----Ejemplito------
 vaciosxD(data, 0.6, F)
 vaciosxD(data, 0.6, T)
}
\author{
\strong{Luis F. Maron}
}
